feat(handler): CIAM handlers. chore(handler): refactoring.
