+++ server/core/httphandler/cors.go
@@ -1,0 +1,15 @@
+package httphandler
+
+import "net/http"
+
+type corsHeaders map[string]string
+
+func (h corsHeaders) setHeaders(header http.Header) {
+	for k, v := range h {
+		header.Set(k, v)
+
+		if k == "Access-Control-Allow-Origin" && (v == "" || v == "'*'") {
+			header.Set(k, "*")
+		}
+	}
+}
